+++ src/utils/createButtonStyle.js
@@ -1,8 +1,4 @@
-import React from 'react';
-
-function Button(props) {
-  const { isDisabled = false, children, onClick, type } = props;
-
+const createButtonStyle = (isDisabled) => {
   const buttonTypeStyle = {
     smallSquare: `h-[32px] rounded-[2px] rounded-[5px] py-2 ${!isDisabled ? 'bg-gradient-to-r from-pointOrange to-pointPink' : 'cursor-not-allowed bg-grayMedium'}`,
     largeSquare: `h-[40px] w-[295px] rounded-[3px] ${!isDisabled ? 'cursor-point bg-gradient-to-r from-pointOrange to-pointPink' : 'cursor-not-allowed bg-grayMedium'}`,
@@ -11,18 +7,7 @@ function Button(props) {
     more: `border-whiteSecondary-500 h-[40px] w-[295px] rounded-[3px] border bg-blackSecondary ${!isDisabled ? '' : 'cursor-not-allowed'}`,
   };
 
-  const buttonTypeName = type ? buttonTypeStyle[type] : '';
+  return buttonTypeStyle;
+};
 
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={isDisabled}
-      className={`cursor-point flex items-center justify-center px-[16px] text-[13px] font-bold ${buttonTypeName}`}
-    >
-      {children}
-    </button>
-  );
-}
-
-export default Button;
+export default createButtonStyle;
